requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1
django-bootstrap-carousel==0.2
django-bootstrap4==2.3.1
django-classy-tags==2.0.0
django-cms==3.8.0
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-garpix-helpers==0.1.3
django-garpix-slider==0.1.14
django-hero-slider==1.3.1
django-heroku==0.3.1
django-humanize==0.1.2
django-hvad==1.8.0
django-import-export==2.4.0
django-include-by-ajax==2.1.0
django-js-asset==1.2.2
django-libs==2.0.3
django-mptt==0.11.0
django-nose==1.4.7
django-polymorphic==3.0.0
django-resized==0.3.11
django-sekizai==2.0.0
django-treebeard==4.3.1
django-uuslug==1.2.0
djangoajax==3.2
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-text-ckeditor==4.0.0
easy-thumbnails==2.7.1
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
humanize==3.1.0
idna==2.8
importlib-metadata==2.1.1
jdcal==1.4.1
MarkupPy==1.14
nose==1.3.7
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
psycopg2==2.8.6
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
six==1.14.0
soupsieve==2.0.1
South==1.0.2
sqlparse==0.3.1
tablib==3.0.0
text-unidecode==1.3
Unidecode==1.1.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0